component split and working nav buttons
